Correçao Bugs

+++ Back/HumanCRM-Api/Dto/ClienteDto.cs
@@ -24,7 +24,7 @@ namespace HumanCRM_Api.Dto
         public int? EstadoCivil { get; set; }
         public int? DDD { get; set; }
         public string? Telefone { get; set; }
-        public int? Celular { get; set; }
+        public long? Celular { get; set; }
         public DateTime? DataNascimento { get; set; }
         public List<ProspeccaoResponseDto>? Prospeccoes { get; set; }
     }
